WIP: getting clang22 working on ARM

+++ src/test/common/Mmap.cpp
@@ -1,5 +1,6 @@
 #include "common/runtime/Mmap.hpp"
 #include <gtest/gtest.h>
+#include <cstdint>
 
 using namespace runtime;
 
@@ -21,7 +22,7 @@ TEST(Mmap, rwCycle) {
 
    unordered_map<int, vector<uint64_t>> m;
    for (size_t i = 0; i < 1000000; i++)
-      m.insert({i, {i}});
+      m.insert({static_cast<int>(i), {i}});
    HashTable<int, uint64_t>::writeBinary("/tmp/mapx", m);
 
    HashTable<int, uint64_t> h("/tmp/mapx");
